upload source code demo.py
Demonstrating the core mechanisms of encryption-based ransomware. In this program, the AES algorithm is used to encrypt files within a specified directory, securing data with strong symmetric encryption. Additionally, RSA is implemented to enhance security by encrypting the AES encryption key itself. This ensures that the private RSA key and the 64-character ciphertext are required to decrypt the files. Once the files are encrypted, the private key and ciphertext are securely transmitted, meaning access to these elements is essential for any decryption attempts.